database is all done concerning car data

+++ Car-Evaluation-App/database/migrations/2024_03_28_071250_car.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->integer('transmission_id');
             $table->integer('drivetype_id');
             $table->integer('fueltype_id');
-            $table->integer('fuelconsumption');
+            $table->float('fuelconsumption');
             $table->integer('kilometers');
             $table->integer('cylinders');
             $table->integer('bodytype_id');
